+++ build.sbt
@@ -28,9 +28,6 @@ inThisBuild(
   )
 )
 
-addCommandAlias("fmt", "all scalafmtSbt scalafmt test:scalafmt")
-addCommandAlias("check", "all scalafmtSbtCheck scalafmtCheck test:scalafmtCheck")
-
 val zioVersion     = "1.0.12"
 val zioJsonVersion = "0.2.0-M2"
 val sttpVersion    = "3.3.16"
@@ -41,92 +38,52 @@ lazy val root =
     .settings(publish / skip := true)
     .aggregate(
       zioLambda,
-      zioLambdaExample,
-      zioRuntime,
-      zioRuntimeLambda,
-      zioRuntimeLambdaExample,
-      zioLambdaShared
+      zioLambdaExample
     )
 
 lazy val zioLambda = module("zio-lambda", "lambda")
   .enablePlugins(BuildInfoPlugin)
+  .enablePlugins(JavaAppPackaging)
   .settings(buildInfoSettings("zio.lambda"))
   .settings(
     stdSettings("zio-lambda"),
-    libraryDependencies ++= Seq(
-      "com.softwaremill.sttp.client3" %% "httpclient-backend" % sttpVersion
-    ),
     testFrameworks := Seq(new TestFramework("zio.test.sbt.ZTestFramework")),
-    assembly / assemblyJarName := "zio-lambda.jar"
+    libraryDependencies ++= Seq(
+      "com.softwaremill.sttp.client3" %% "okhttp-backend" % sttpVersion
+    )
   )
-  .dependsOn(zioRuntime)
-  .dependsOn(zioLambdaShared)
+  .settings(
+    topLevelDirectory := None,
+    Universal / mappings ++= Seq(file("bootstrap") -> "bootstrap"),
+    Compile / mainClass := Some("zio.lambda.ZRuntimeApp")
+  )
 
 lazy val zioLambdaExample = module("zio-lambda-example", "lambda-example")
-  .enablePlugins(NativeImagePlugin)
   .enablePlugins(BuildInfoPlugin)
+  .enablePlugins(GraalVMNativeImagePlugin)
   .settings(buildInfoSettings("zio.lambda.example"))
   .settings(
+    name := "zio-lambda-example",
     stdSettings("zio-lambda-example"),
     assembly / assemblyJarName := "zio-lambda-example.jar",
-    nativeImageOptions ++= List("--no-fallback", "--enable-http"),
-    nativeImageOptions += s"-H:ReflectionConfigurationFiles=${target.value / "native-image-configs" / "reflect-config.json"}",
-    nativeImageOptions += s"-H:ConfigurationFileDirectories=${target.value / "native-image-configs"}",
-    nativeImageOptions += "-H:+JNI",
-    nativeImageOptions += "-H:+AllowIncompleteClasspath"
+    GraalVMNativeImage / mainClass := Some("zio.lambda.example.SimpleHandler"),
+    graalVMNativeImageOptions := Seq(
+      "--verbose",
+      "--no-fallback",
+      "--install-exit-handlers",
+      "--enable-http",
+      "--allow-incomplete-classpath",
+      "--report-unsupported-elements-at-runtime",
+      "-H:+StaticExecutableWithDynamicLibC",
+      "-H:+RemoveSaturatedTypeFlows"
+    )
   )
   .dependsOn(zioLambda)
-
-lazy val zioRuntime = module("zio-runtime", "runtime")
-  .enablePlugins(BuildInfoPlugin)
-  .settings(buildInfoSettings("zio.runtime"))
-  .settings(stdSettings("zio-runtime"))
-  .settings(
-    libraryDependencies ++= Seq(
-      "com.softwaremill.sttp.client3" %% "httpclient-backend" % "3.3.15"
-    ),
-    testFrameworks := Seq(new TestFramework("zio.test.sbt.ZTestFramework")),
-    assembly / assemblyJarName := "zio-runtime.jar",
-    assembly / assemblyMergeStrategy := {
-      case "META-INF/io.netty.versions.properties" =>
-        MergeStrategy.concat
-      case x => (assembly / assemblyMergeStrategy).value(x)
-    }
-  )
-  .dependsOn(zioLambdaShared)
-  .dependsOn(zioRuntimeLambda)
-
-lazy val zioRuntimeLambda = module("zio-runtime-lambda", "runtime-lambda")
-  .enablePlugins(BuildInfoPlugin)
-  .settings(buildInfoSettings("zio.runtime.lambda"))
-  .settings(stdSettings("zio-runtime"))
-  .settings(testFrameworks := Seq(new TestFramework("zio.test.sbt.ZTestFramework")))
-  .dependsOn(zioLambdaShared)
-
-lazy val zioRuntimeLambdaExample = module("zio-runtime-lambda-example", "runtime-lambda-example")
-  .enablePlugins(NativeImagePlugin)
-  .enablePlugins(BuildInfoPlugin)
-  .settings(buildInfoSettings("zio.runtime.lambda.example"))
-  .settings(
-    stdSettings("zio-runtime-lambda-example"),
-    assembly / assemblyJarName := "zio-runtime-lambda-example.jar"
-  )
-  .dependsOn(zioRuntimeLambda)
-
-lazy val zioLambdaShared = module("zio-lambda-shared", "shared")
-  .enablePlugins(BuildInfoPlugin)
-  .settings(buildInfoSettings("zio.lambda.shared"))
-  .settings(stdSettings("zio-lambda-shared"))
-  .settings(testFrameworks := Seq(new TestFramework("zio.test.sbt.ZTestFramework")))
 
 def module(moduleName: String, fileName: String): Project =
   Project(moduleName, file(fileName))
     .settings(stdSettings(moduleName))
     .settings(
-      assembly / assemblyExcludedJars := {
-        val cp = (assembly / fullClasspath).value
-        cp filter (_.data.getName.contains("scalaz-core"))
-      },
       libraryDependencies ++= Seq(
         "dev.zio" %% "zio"          % zioVersion,
         "dev.zio" %% "zio-test"     % zioVersion % "test",
@@ -153,3 +110,6 @@ lazy val docs = project
   )
   .dependsOn(zioLambda)
   .enablePlugins(MdocPlugin, DocusaurusPlugin, ScalaUnidocPlugin)
+
+addCommandAlias("fmt", "all scalafmtSbt scalafmt test:scalafmt")
+addCommandAlias("check", "all scalafmtSbtCheck scalafmtCheck test:scalafmtCheck")
