+++ lambda/src/main/scala/zio/lambda/InvocationError.scala
@@ -1,0 +1,6 @@
+package zio.lambda
+
+final case class InvocationError(
+  requestId: InvocationRequest.Id,
+  errorResponse: InvocationErrorResponse
+)
