+++ lambda/src/main/scala/zio/lambda/InvocationResponse.scala
@@ -1,0 +1,3 @@
+package zio.lambda
+
+final case class InvocationResponse(requestId: InvocationRequest.Id, payload: String)
