+++ project/plugins.sbt
@@ -1,10 +1,12 @@
-addSbtPlugin("org.scalameta"      % "sbt-scalafmt"     % "2.4.4")
-addSbtPlugin("pl.project13.scala" % "sbt-jmh"          % "0.4.3")
-addSbtPlugin("com.eed3si9n"       % "sbt-buildinfo"    % "0.10.0")
-addSbtPlugin("org.scoverage"      % "sbt-scoverage"    % "1.9.2")
-addSbtPlugin("org.scalameta"      % "sbt-mdoc"         % "2.2.24")
-addSbtPlugin("ch.epfl.scala"      % "sbt-bloop"        % "1.4.11")
-addSbtPlugin("com.eed3si9n"       % "sbt-unidoc"       % "0.4.3")
-addSbtPlugin("com.github.sbt"     % "sbt-ci-release"   % "1.5.10")
-addSbtPlugin("com.eed3si9n"       % "sbt-assembly"     % "0.14.6")
-addSbtPlugin("org.scalameta"      % "sbt-native-image" % "0.3.2")
+addSbtPlugin("org.scalameta"      % "sbt-scalafmt"        % "2.4.4")
+addSbtPlugin("pl.project13.scala" % "sbt-jmh"             % "0.4.3")
+addSbtPlugin("com.eed3si9n"       % "sbt-buildinfo"       % "0.10.0")
+addSbtPlugin("org.scoverage"      % "sbt-scoverage"       % "1.9.2")
+addSbtPlugin("org.scalameta"      % "sbt-mdoc"            % "2.2.24")
+addSbtPlugin("ch.epfl.scala"      % "sbt-bloop"           % "1.4.11")
+addSbtPlugin("com.eed3si9n"       % "sbt-unidoc"          % "0.4.3")
+addSbtPlugin("com.github.sbt"     % "sbt-ci-release"      % "1.5.10")
+addSbtPlugin("com.eed3si9n"       % "sbt-assembly"        % "0.14.6")
+addSbtPlugin("org.scalameta"      % "sbt-native-image"    % "0.3.2")
+addSbtPlugin("com.typesafe.sbt"   % "sbt-native-packager" % "1.8.1")
+addSbtPlugin("com.eed3si9n"       % "sbt-assembly"        % "0.14.6")
