--- runtime-lambda-example/src/main/scala/zio/runtime/lambda/example/CustomEvent.scala
@@ -1,9 +1,0 @@
-package zio.runtime.lambda.example
-
-import zio.json._
-
-final case class CustomEvent(message: String)
-
-object CustomEvent {
-  implicit val decoder: JsonDecoder[CustomEvent] = DeriveJsonDecoder.gen[CustomEvent]
-}
