--- runtime-lambda-example/src/main/scala/zio/runtime/lambda/example/CustomResponse.scala
@@ -1,9 +1,0 @@
-package zio.runtime.lambda.example
-
-import zio.json._
-
-final case class CustomResponse(message: String)
-
-object CustomResponse {
-  implicit val encoder: JsonEncoder[CustomResponse] = DeriveJsonEncoder.gen[CustomResponse]
-}
