--- runtime-lambda-example/src/main/scala/zio/runtime/lambda/example/SimpleHandler.scala
@@ -1,14 +1,0 @@
-package zio.runtime.lambda.example
-
-import zio._
-import zio.console._
-import zio.runtime.lambda.ZLambda
-
-final class SimpleHandler extends ZLambda[CustomEvent, CustomResponse] {
-
-  override def handle(request: CustomEvent): RIO[ZEnv, CustomResponse] =
-    for {
-      _ <- putStrLn(request.message)
-    } yield CustomResponse("Handler ran successfully")
-
-}
