--- runtime-lambda/src/main/scala/zio/runtime/lambda/ZLambda.scala
@@ -1,47 +1,0 @@
-package zio.runtime.lambda
-
-import zio._
-import zio.json._
-
-/**
- * Implementation example:
- *
- * {{{
- * object MyLambda extends ZLambda[MyPayload, MyResponse] {
- *  def handle(request: MyPayload): ZIO[ZEnv, Throwable, MyResponse] = ???
- * }
- *
- * final case class MyPayload(value: String)
- *
- * object MyPayload {
- *  import zio.json.JsonDecoder
- *  import zio.json.DeriveJsonDecoder
- *
- *  implicit val jsonDecoder: JsonDecoder[MyPayload] = DeriveJsonDecoder.gen
- * }
- *
- * final case class MyResponse(value: String)
- *
- * object MyResponse {
- *  import zio.json.JsonEncoder
- *  import zio.json.DeriveJsonEncoder
- *
- *  implicit val jsonEncoder: JsonEncoder[MyResponse] = DeriveJsonEncoder.gen
- * }
- * }}}
- */
-abstract class ZLambda[E, A](
-  implicit val eventDecoder: JsonDecoder[E],
-  implicit val responseEncoder: JsonEncoder[A]
-) {
-
-  def handle(event: E): RIO[ZEnv, A]
-
-  final def run(json: String): RIO[ZEnv, String] =
-    eventDecoder.decodeJson(json) match {
-      case Left(errorMessage) =>
-        ZIO.fail(new Throwable(s"Error decoding json. Json=$json, Error$errorMessage"))
-
-      case Right(event) => handle(event).map(_.toJson)
-    }
-}
